.github/actions/project_update/requirements.txt
PyGithub
requests
